tainted: mirdeep2: Add test (tutorial)
Since the package definition had to practically re-implement the entire
installation procedure, the tutorial is run post-install (using the
patched and wrapped outputs).

The tutorial script 'handles' errors by printing an error message (incl.
the return value of the failed command) to stdout [sic]. Then, it
continues running as if nothing happens (potentially causing and
reporting further errors) and always returns zero in the end. Thus,
stdout is scanned for such error messages to determine failure or
success of the test.
Additional error may occur that are not handled by the tutorial script.
These result in error messages to stderr. Thus, stderr is combined with
stdout before scanning for error messages.
Scanning the tutorial script's output for error messages 'consumes' it,
such that it won't show in the Guix build log anymore. This, in case of
an error, understanding its nature would be harder than necessary.
Once option would be to report the actual error message found, but
additional context might be required. Instead of hard coding a fixed
number of pre-error lines to include in the build log (by replacing
`grep`'s `-q` flag by a `-B` one), the entire output of the tutorial
script (up to the error) is always included in the build log: This is
achieved by copying the (combined) output to stdout *and* stderr. This
way, the stdout copy can be consumed for the downstream checks but the
stderr one remains visible in the log.